requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Flask-Session==0.5.0
openpyxl==3.1.2
pandas==2.1.4
chardet==5.2.0
